+++ divvy/compute.py
@@ -12,7 +12,7 @@ from yaml import SafeLoader
 from distutils.dir_util import copy_tree
 
 # from attmap import PathExAttMap
-from ubiquerg import is_writable
+from ubiquerg import is_writable, VersionInHelpParser
 import yacman
 from collections import OrderedDict
 
@@ -45,36 +45,25 @@ class ComputingConfiguration(yacman.YacAttMap):
         `DIVCFG` file)
     """
 
-    def __init__(self, entries=None, filepath=None, 
-                config_file=None,  # for backwards compatibility with peppy 0.22
-                no_env_error=None,  # for backwards compatibility with peppy 0.22
-                no_compute_exception=None): # for backwards compatibility with peppy 0.22
+    def __init__(self, entries=None, filepath=None): # for backwards compatibility with peppy 0.22
 
-        if no_env_error:
-            _LOGGER.debug("The no_env_error argument has been deprecated. It will be removed in the next version of divvy")
-
-        if no_compute_exception:
-            _LOGGER.debug("The no_compute_exception argument has been deprecated. It will be removed in the next version of divvy")
-
-        if config_file:  # for backwards compatibility with peppy 0.22 (remove later)
-            _LOGGER.debug("The config_file argument has renamed filepath.")
-            filepath = select_divvy_config(config_file)
-
-        if not entries and not filepath and not config_file:
+        if not entries and not filepath:
             # Handle the case of an empty one, when we'll use the default
             filepath = select_divvy_config(None)
 
         super(ComputingConfiguration, self).__init__(entries, filepath)
 
         if not hasattr(self, "compute_packages"):
-            raise Exception("Your divvy config file is not in divvy config format (it"
-            " lacks a compute_packages section): '{}'".format(filepath))
+            raise Exception(
+                "Your divvy config file is not in divvy config format "
+                "(it lacks a compute_packages section): '{}'".format(filepath))
             # We require that compute_packages be present, even if empty
             self.compute_packages = {}
 
         # Initialize default compute settings.
         _LOGGER.debug("Establishing project compute settings")
         self.compute = None
+        self.setdefault("adapters", None)
         self.activate_package(DEFAULT_COMPUTE_RESOURCES_NAME)
         self.config_file = self._file_path
 
@@ -87,6 +76,8 @@ class ComputingConfiguration(yacman.YacAttMap):
             print(pkg)
             destfile = os.path.join(filedir, os.path.basename(pkg.submission_template))
             shutil.copyfile(pkg.submission_template, destfile)
+
+
 
     @property
     def compute_env_var(self):
@@ -232,26 +223,99 @@ class ComputingConfiguration(yacman.YacAttMap):
         self.update(entries)
         return True
 
+    def get_adapters(self):
+        """
+        Get current adapters, if defined.
+
+        Adapters are sourced from the 'adapters' section in the root of the
+        divvy configuration file and updated with an active compute
+        package-specific set of adapters, if any defined in 'adapters' section
+        under currently active compute package.
+
+        :return yacman.YacAttMap: current adapters mapping
+        """
+        adapters = yacman.YacAttMap()
+        if "adapters" in self and self.adapters is not None:
+            adapters.update(self.adapters)
+        if "compute" in self and "adapters" in self.compute:
+            adapters.update(self.compute.adapters)
+        if not adapters:
+            _LOGGER.debug("No adapters determined in divvy configuration file.")
+        return adapters
+
+    def submit(self, output_path, extra_vars=None):
+        if not output_path:
+            import tempfile
+            with tempfile.NamedTemporaryFile() as temp:
+                _LOGGER.info("No file provided; using temp file: '{}'".format(temp.name))
+                self.submit(temp.name, extra_vars)
+        else:
+            script = self.write_script(output_path, extra_vars)
+            submission_command = "{submit_command} {submit_script}".format(submit_command=self.compute.submission_command, 
+                submit_script=script)
+            _LOGGER.info(submission_command)
+            os.system(submission_command)
+
     def write_script(self, output_path, extra_vars=None):
         """
         Given currently active settings, populate the active template to write a
-         submission script.
+         submission script. Additionally use the current adapters to adjust
+         the select of the provided variables
 
         :param str output_path: Path to file to write as submission script
-        :param Iterable[Mapping] extra_vars: A list of Dict objects with key-value pairs
-            with which to populate template fields. These will override any
-            values in the currently active compute package.
+        :param Iterable[Mapping] extra_vars: A list of Dict objects with
+            key-value pairs with which to populate template fields. These will
+            override any values in the currently active compute package.
         :return str: Path to the submission script file
         """
+
+        def _get_from_dict(map, attrs):
+            """
+            Get value from a possibly mapping using a list of its attributes
+
+            :param collections.Mapping map: mapping to retrieve values from
+            :param Iterable[str] attrs: a list of attributes
+            :return: value found in the for the requested attribute or
+                None if one of the keys does not exist
+            """
+            for a in attrs:
+                try:
+                    map = map[a]
+                except KeyError:
+                    return None
+            return map
+
         from copy import deepcopy
         variables = deepcopy(self.compute)
+        _LOGGER.debug("Extra vars: {}".format(extra_vars))
         if extra_vars:
             if not isinstance(extra_vars, list):
                 extra_vars = [extra_vars]
-            for kvs in reversed(extra_vars):
-                variables.update(kvs)
-        _LOGGER.info("Writing script to {}".format(os.path.abspath(output_path)))
-        _LOGGER.debug("Submission template: {}".format(self.compute.submission_template))
+            adapters = self.get_adapters()
+            exclude = set()
+            if adapters:
+                # apply adapted values first and keep track of
+                # which of extra_vars were used
+                for n, v in adapters.items():
+                    split_v = v.split(".")
+                    namespace = split_v[0]
+                    for extra_var in reversed(extra_vars):
+                        if len(extra_var) > 0 and namespace in list(extra_var.keys())[0]:
+                            exclude.add(namespace)
+                            var = _get_from_dict(extra_var, split_v)
+                            if var is not None:
+                                variables[n] = var
+                                _LOGGER.debug("adapted {}: ({}={})".
+                                              format(n, ".".join(split_v), var))
+            for extra_var in reversed(extra_vars):
+                # then update variables with the rest of the extra_vars
+                if len(extra_var) > 0 and list(extra_var.keys())[0] not in exclude:
+                    variables.update(extra_var)
+        _LOGGER.debug("Submission template: {}".
+                      format(self.compute.submission_template))
+        if output_path:
+            _LOGGER.info("Writing script to {}".
+                         format(os.path.abspath(output_path)))
         return write_submit_script(output_path, self.template(), variables)
 
     def _handle_missing_env_attrs(self, config_file, when_missing):
@@ -270,20 +334,24 @@ class ComputingConfiguration(yacman.YacAttMap):
 
 
 def select_divvy_config(filepath):
+    """
+    Selects the divvy config file path to load.
+
+    This uses a priority ordering to first choose a config file path if
+    it's given, but if not, then look in a priority list of environment
+    variables and choose the first available file path to return. If none of
+    these options succeed, the default config path will be returned.
+
+    :param str | NoneType filepath: direct file path specification
+    :return str: path to the config file to read
+    """
     divcfg = yacman.select_config(
-        filepath,
-        COMPUTE_SETTINGS_VARNAME,
+        config_filepath=filepath,
+        config_env_vars=COMPUTE_SETTINGS_VARNAME,
         default_config_filepath=DEFAULT_CONFIG_FILEPATH,
         check_exist=True)
     _LOGGER.debug("Selected divvy config: {}".format(divcfg))
     return divcfg
-
-
-class _VersionInHelpParser(argparse.ArgumentParser):
-    def format_help(self):
-        """ Add version information to help text. """
-        return "version: {}\n".format(__version__) + \
-               super(_VersionInHelpParser, self).format_help()
 
 
 def divvy_init(config_path, template_config_path):
@@ -322,18 +390,13 @@ def build_argparser():
     :return argparse.ArgumentParser
     """
 
-    banner = "%(prog)s - write compute job scripts that can be submitted to any computing resource"
+    banner = "%(prog)s - write compute job scripts that can be submitted to " \
+             "any computing resource"
     additional_description = "\nhttps://divvy.databio.org"
 
-    parser = _VersionInHelpParser(
-            prog="divvy",
-            description=banner,
-            epilog=additional_description)
-
-    parser.add_argument(
-            "-V", "--version",
-            action="version",
-            version="%(prog)s {v}".format(v=__version__))
+    parser = VersionInHelpParser(prog="divvy", description=banner,
+                                 epilog=additional_description,
+                                 version=__version__)
 
     subparsers = parser.add_subparsers(dest="command") 
 
@@ -344,31 +407,46 @@ def build_argparser():
     subparser_messages = {
         "init": "Initialize a new divvy config file",
         "list": "List available compute packages",
-        "write": "Write a job script"
+        "write": "Write a job script",
+        "submit": "Write and then submit a job script"
     }
 
     sps = {}
     for cmd, desc in subparser_messages.items():
         sps[cmd] = add_subparser(cmd, desc)
-        sps[cmd].add_argument(
-            "-c", "--config", required=(cmd == "init"),
+        # sps[cmd].add_argument(
+        #     "config", nargs="?", default=None,
+        #     help="Divvy configuration file.")
+
+    for sp in [sps["list"], sps["write"], sps["submit"]]:
+        sp.add_argument(
+            "config", nargs="?", default=None,
             help="Divvy configuration file.")
 
-    sps["write"].add_argument(
-            "-s", "--settings",
-            help="YAML file with job settings to populate the template.")    
+    sps["init"].add_argument(
+            "config", default=None,
+            help="Divvy configuration file.")
 
-    sps["write"].add_argument(
-            "-p", "--package", default=DEFAULT_COMPUTE_RESOURCES_NAME,
-            help="Select from available compute packages.")
+    for sp in [sps["write"], sps["submit"]]:
+        sp.add_argument(
+                "-s", "--settings",
+                help="YAML file with job settings to populate the template")
 
-    # sps["write"].add_argument(
-    #         "-t", "--template",
-    #         help="Provide a template file (not yet implemented).")
+        sp.add_argument(
+                "-p", "--package", default=DEFAULT_COMPUTE_RESOURCES_NAME,
+                help="Select from available compute packages")
 
-    sps["write"].add_argument(
-            "-o", "--outfile", required=True,
-            help="Output filepath")
+        sp.add_argument(
+                "-c", "--compute", nargs="+", default=None,
+                help="Extra key=value variable pairs")
+
+        # sp.add_argument(
+        #         "-t", "--template",
+        #         help="Provide a template file (not yet implemented).")
+
+        sp.add_argument(
+                "-o", "--outfile", required=False, default=None,
+                help="Output filepath")
 
     return parser
 
@@ -377,9 +455,11 @@ def main():
     """ Primary workflow """
 
     parser = logmuse.add_logging_options(build_argparser())
-    args, remaining_args = parser.parse_known_args()
+    # args, remaining_args = parser.parse_known_args()
+    args = parser.parse_args()
+
     logger_kwargs = {"level": args.verbosity, "devmode": args.logdev}
-    logmuse.init_logger(name="yacman", **logger_kwargs)
+    logmuse.init_logger("yacman", **logger_kwargs)
     global _LOGGER
     _LOGGER = logmuse.logger_via_cli(args)
 
@@ -387,10 +467,6 @@ def main():
         parser.print_help()
         _LOGGER.error("No command given")
         sys.exit(1)
-    # Any non-divvy arguments will be passed along as key-value pairs
-    # that can be used to populate the template.
-    keys = [str.replace(x, "--", "") for x in remaining_args[::2]]
-    cli_vars = dict(zip(keys, remaining_args[1::2]))
 
     if args.command == "init":
         divcfg = args.config
@@ -399,6 +475,7 @@ def main():
         divvy_init(divcfg, DEFAULT_CONFIG_FILEPATH)
         sys.exit(0)      
 
+    _LOGGER.debug("Divvy config: {}".format(args.config))
     divcfg = select_divvy_config(args.config)
     _LOGGER.info("Using divvy config: {}".format(divcfg))
     dcc = ComputingConfiguration(filepath=divcfg)
@@ -410,17 +487,32 @@ def main():
         print("{}".format("\n".join(dcc.list_compute_packages())))
         sys.exit(1)
 
-
-    try:
-        dcc.activate_package(args.package)
-    except AttributeError:
-        parser.print_help(sys.stderr)
-        sys.exit(1)
-
-    if args.settings:
-        _LOGGER.info("Loading settings file: %s", args.settings)
-        with open(args.settings, 'r') as f:
-            vars_groups = [cli_vars, yaml.load(f, SafeLoader)]
+    # Any non-divvy arguments will be passed along as key-value pairs
+    # that can be used to populate the template.
+    # keys = [str.replace(x, "--", "") for x in remaining_args[::2]]
+    # cli_vars = dict(zip(keys, remaining_args[1::2]))
+    if args.compute:
+        cli_vars = {y[0]:y[1] for y in [x.split("=") for x in args.compute]}
     else:
-        vars_groups = [cli_vars]
-    dcc.write_script(args.outfile, vars_groups)
+        cli_vars = {}
+
+
+    if args.command == "write" or args.command == "submit":
+        try:
+            dcc.activate_package(args.package)
+        except AttributeError:
+            parser.print_help(sys.stderr)
+            sys.exit(1)
+
+        if args.settings:
+            _LOGGER.info("Loading settings file: %s", args.settings)
+            with open(args.settings, 'r') as f:
+                vars_groups = [cli_vars, yaml.load(f, SafeLoader)]
+        else:
+            vars_groups = [cli_vars]
+
+        _LOGGER.debug(vars_groups)
+        if args.command == "write":
+            dcc.write_script(args.outfile, vars_groups)
+        elif args.command == "submit":
+            dcc.submit(args.outfile, vars_groups)