+++ docs/changelog.md
@@ -1,4 +1,22 @@
+This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html) and [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) format. 
+
 # Changelog
+
+## [0.5.0] -- 2020-05-18
+### Added
+- adapters support, see [#47](https://github.com/pepkit/divvy/issues/47) for detailed explanation
+- `select_divvy_config` function
+- `divvy submit` command
+- added new `--compute` CLI argument
+- `divvy write` can now be run without an outfile, which just prints the template to stdout.
+
+### Changed
+- Instead of passing extra variables as CLI args, you now must explicitly pass them to the `--compute` arg.
+- The config file is now passed with a positional argument, instead of with `--config`.
+- Made all one-char CLI args lowercase (`-P` to `-p`, `-S` to `-s`).
+
+### Removed
+- `config_file`, `no_env_error` and `no_compute_exception` from `ComputingConfiguration` class constructor
 
 ## [0.4.1] -- 2020-03-20
 ### Fixed